# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "pegasus",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/bots": {
      "get": {
        "tags": [
          "Bot",
          "bot_controller"
        ],
        "description": "Get all online bots including all user info.",
        "operationId": "index",
        "requestBody": {
          "description": "Payload to request bots based on the hotel",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BotsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "List of available bots",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailableBots"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/bots/{ticket}": {
      "get": {
        "tags": [
          "Bot",
          "bot_controller"
        ],
        "description": "Get a single bot's information.",
        "operationId": "show",
        "parameters": [
          {
            "name": "ticket",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Payload to request a single based on the hotel",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShowBotRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The requested bot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailableBots"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "AvailableBots": {
        "type": "object",
        "properties": {
          "bots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BotInfo"
            },
            "nullable": true
          }
        }
      },
      "BotInfo": {
        "type": "object",
        "required": [
          "user_id",
          "username",
          "motto",
          "figure",
          "gender",
          "sso_ticket"
        ],
        "properties": {
          "figure": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "motto": {
            "type": "string"
          },
          "sso_ticket": {
            "type": "string"
          },
          "user_id": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "username": {
            "type": "string"
          }
        }
      },
      "BotsRequest": {
        "type": "object",
        "required": [
          "hotel"
        ],
        "properties": {
          "hotel": {
            "type": "string"
          }
        }
      },
      "BulkUpdateResponse": {
        "type": "object",
        "required": [
          "task_id"
        ],
        "properties": {
          "task_id": {
            "type": "string"
          }
        }
      },
      "ShowBotRequest": {
        "type": "object",
        "required": [
          "hotel"
        ],
        "properties": {
          "hotel": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-auth-token"
      }
    }
  }
}